Move tests that belong inside insied.
